Working version, library specific
The main file works as intended but the full search logic only makes sense with the way I logically organize my library (artist>album>songs) for handling tag discrepancies. In future I'd like to rewrite all of that logic because it's not very clean, but for now this will do!